third_party/WebKit/LayoutTests/typedcssom/stylevalue-normalization/normalize-resource-expected.txt

This is a testharness.js-based test.
PASS Normalizing a valid <url> returns a CSSURLImageValue that eventually loads
PASS Normalizing a bad <url> returns a CSSURLImageValue in error state
FAIL Normalizing a <gradient> returns a CSSImageValue assert_equals: expected "CSSImageValue" but got "CSSStyleValue"
Harness: the test ran to completion.
